Add useScaffoldEventRead hook and show withdraw events
useScaffoldEventRead hook allows to get the events from a contract.

Only contractName and eventName are mandatory.

You can choose if you want to get the block, transaction or receipt data with the blockData, transactionData and receiptData flags. Default to false to avoid RPC calls.

fromBlock default value is set to 1000 blocks before the current block. If you want to get all the events from the contract, set it to the block number when the contract was deployed.

Use filters to filters the events by the indexed parameters.

Example usage:

  const events = useScaffoldEventRead({
    contractName: "YourContract",
    eventName: "Withdraw",
    fromBlock: 0,
    blockData: true,
    transactionData: true,
    receiptData: true,
    filters: {
    	to: "0x5dCb5f4F39Caa6Ca25380cfc42280330b49d3c93",
    }
  });